+++ src/Database/PostgreSQL/PQTypes/Model/Check.hs
@@ -15,8 +15,8 @@ import Prelude
 data Check = Check {
   chkName      :: RawSQL ()
 , chkCondition :: RawSQL ()
-, chkValidated :: Bool -- ^ Set to 'True' if check is created as NOT VALID and
-                       -- not validated afterwards.
+, chkValidated :: Bool -- ^ Set to 'False' if check is created as NOT VALID and
+                       -- left in such state (for whatever reason).
 } deriving (Eq, Ord, Show)
 
 tblCheck :: Check